feat: migrate from embedded Postgres to SQLite (#40)
* feat: migrate from embedded Postgres to SQLite

Replaces embedded-postgres + postgres.js with better-sqlite3 for a
simpler, faster, zero-process-management local database. The database
is now a single file at ~/.arkeon-wiki/data/arke.db — no child
processes, ports, passwords, or orphan cleanup needed.

All 58 tests pass (28 unit + 30 e2e) against SQLite.


* fix: address PR review — transaction mutex, shared db, dep cleanup

- Add transaction queue to serialize concurrent withTransaction calls,
  preventing SQLITE_BUSY on overlapping async transactions
- Make migrate.ts use shared initDb() instead of creating a separate
  connection with duplicated pragma setup
- Remove unused dependencies (meilisearch, openai, @aws-sdk/client-s3,
  js-yaml, @types/js-yaml)


---------

+++ packages/arkeon/src/server/routes/spaces.ts
@@ -41,7 +41,7 @@ spacesRouter.get("/", async (c) => {
   const sql = createSql();
   const spaces = await sql`
     SELECT s.id, s.name, s.watch_dir, s.created_at,
-           COUNT(e.id)::int AS entity_count
+           COUNT(e.id) AS entity_count
     FROM spaces s
     LEFT JOIN entities e ON e.space_id = s.id
     GROUP BY s.id
@@ -56,7 +56,7 @@ spacesRouter.get("/:id", async (c) => {
   const sql = createSql();
   const rows = await sql`
     SELECT s.id, s.name, s.watch_dir, s.created_at,
-           COUNT(e.id)::int AS entity_count
+           COUNT(e.id) AS entity_count
     FROM spaces s
     LEFT JOIN entities e ON e.space_id = s.id
     WHERE s.id = ${c.req.param("id")}
